runner: support awaitable config and basic logic

+++ lisa/schema.py
@@ -1430,7 +1430,7 @@ class Runbook:
     tags: Optional[List[str]] = None
     concurrency: int = 1
     # minutes to wait for resource
-    wait_resource_timeout: int = 5
+    wait_resource_timeout: float = 5
     include: Optional[List[Include]] = field(default=None)
     extension: Optional[List[Union[str, Extension]]] = field(default=None)
     variable: Optional[List[Variable]] = field(default=None)
